requirements.txt
langchain
langchain[openai]
faiss-cpu
pydantic
fastapi
uvicorn
python-dotenv
pytest
pytest-cov
sentence-transformers
faiss-cpu
numpy